Move duplicated MetricProvider into common include file. Font assert, uncond return, and remove unreachable.
